Rebuild service state during sync (#1)
Services occasionally fail their Consul TCP health check despite being
able to telnet to their ports.  When in this state, Consul displays
their IP as their Docker network's gateway IP, which is incorrect.

When debugging Registrator, Bridge.service[#]'s Service.IP was observed
to be "" while Docker had a value for the IP. It is currently unknown
how Registrator got into this state.  Nevertheless, because the service
information is cached, the service is never re-registered with Consul
using the correct IP.

- Bumped VERSION.

- Added debugging run configuration.

- Updated Bridge.Sync to rebuild service state.

- Always log when syncing.